refactor(views): rename BookView → Renderer + FixedBookView → FixedRenderer
Pure rename across views and consumers. No behavior change.

- BookView interface → Renderer (file rename + RendererHost rename)
- FixedBookView → FixedRenderer (file rename)
- updateBookView → updateRenderer (EpubNavigator method)
- bookViewPosition → rendererPosition (EpubNavigator local)
- IFrameAttributes import path tidied to "navigator/types"
