+++ src/views/FixedRenderer.ts
@@ -17,13 +17,13 @@
  * Licensed to: CAST under one or more contributor license agreements.
  */
 
-import { IFrameAttributes } from "../navigator/EpubNavigator";
-import BookView, { BookViewHost } from "./BookView";
+import { IFrameAttributes } from "../navigator/types";
+import Renderer, { RendererHost } from "./Renderer";
 import * as HTMLUtilities from "../utils/HTMLUtilities";
 
-export default class FixedBookView implements BookView {
+export default class FixedRenderer implements Renderer {
   layout = "fixed";
-  host: BookViewHost;
+  host: RendererHost;
   name: string;
   label: string;
   iframe: HTMLIFrameElement;
